+++ packages/webcomponents/CHANGELOG.md
@@ -1,5 +1,12 @@
 ### Changelog
 
+## 5.4.4
+
+### Patch Changes
+
+- cff6ebe: - Updated name of column on `justifi-terminals-list` from `Provider ID` to `Device ID`
+  - Fixed bug in `justifi-order-terminals` where header for address section was not rendered correctly.
+
 ## 5.4.3
 
 ### Patch Changes